Extract the Runner package
